Fix yarn lock & tests (also improve resolution)
The last commit broke all the tests since we're now doing proper value set resolution.  Fixed that and updated tests to load dependency definitions first.

This also updates the importer to all a Specifications to be passed in, which will be used for resolving definitions (in addition to the preprocessor).