WIP: Support constraining type to a choice
This update reflects *some* support of the new grammar to allow constraining a field to a choice of sub-types (or constraining an existing value choice to a small subset).

This is incomplete as the model hasn't been updated to support type constraints to choices.  Given the ramifications to the toolchain, this likely requires more thought before moving forward.