Content Profile paths must use the effectiveIdentifier
After discussion w/ Mark, we determined that CPs should always use the "effective identifiers" in the paths.  For example, if field Foo is constrained to SonOfFoo, then the CP path should use SonOfFoo.

In addition fixed an issue where the ChoiceValue options were being iterated as identifiers when they are really Values (that have identifiers).